refactor(prefill): move layer-completion internal API headers to disaggregation/
Relocate the new layer_completion_{message,queue,reorder_buffer,router}.hpp
headers from tt_metal/api/internal/service/ to a dedicated
tt_metal/api/internal/disaggregation/ directory and update all includes.
inter_process_counter_channel.hpp stays in service/ (pre-existing infra).
