+++ tests/tt_metal/distributed/test_layer_completion_queue.cpp
@@ -7,8 +7,8 @@
 #include <cstdio>
 #include <string>
 
-#include <internal/service/layer_completion_message.hpp>
-#include <internal/service/layer_completion_queue.hpp>
+#include <internal/disaggregation/layer_completion_message.hpp>
+#include <internal/disaggregation/layer_completion_queue.hpp>
 
 namespace tt::tt_metal::distributed::test {
