+++ tests/tt_metal/distributed/test_layer_completion_queue_mpsc.cpp
@@ -10,8 +10,8 @@
 #include <unordered_set>
 #include <vector>
 
-#include <internal/service/layer_completion_message.hpp>
-#include <internal/service/layer_completion_queue.hpp>
+#include <internal/disaggregation/layer_completion_message.hpp>
+#include <internal/disaggregation/layer_completion_queue.hpp>
 
 namespace tt::tt_metal::distributed::test {
