+++ tt_metal/api/internal/disaggregation/layer_completion_reorder_buffer.hpp
@@ -16,7 +16,7 @@
 #include <map>
 #include <vector>
 
-#include <internal/service/layer_completion_message.hpp>
+#include <internal/disaggregation/layer_completion_message.hpp>
 
 namespace tt::tt_metal::distributed {
