+++ tt_metal/distributed/layer_completion_ring_layout.hpp
@@ -20,7 +20,7 @@
 #include <cstddef>
 #include <cstdint>
 
-#include <internal/service/layer_completion_message.hpp>
+#include <internal/disaggregation/layer_completion_message.hpp>
 
 namespace tt::tt_metal::distributed {
